Move to CMake, fix bag, add version text
The version text is drawn in the corner of the screen
Bag RNG now uses srand() and rand()
Clean up some of the code
