+++ include/3dstris/version.hpp.in
@@ -1,0 +1,6 @@
+#pragma once
+
+#define _3DSTRIS_VERSION "@3dstris_VERSION@"
+#define _3DSTRIS_VERSION_MAJOR "@3dstris_VERSION_MAJOR@"
+#define _3DSTRIS_VERSION_MINOR "@3dstris_VERSION_MINOR@"
+#define _3DSTRIS_VERSION_PATCH "@3dstris_VERSION_PATCH@"
